Posició jugadors en panalla, correcció posició random inicial, i turn plantejat

+++ Platilla_3Projects/GameLib/PlayerInfo.h
@@ -6,7 +6,7 @@ class PlayerInfo
 {
 public:
 	std::string nickname;
-	sf::Vector2i position;
+	sf::Vector2f position;
 	sf::Color color;
 	std::vector<card> playerCards;
